[FIX] account_bank_statement_import_camt: check exchange rate
The amount of transactions of imported CAMT053 was sometimes inverted
depending on the format used by the bank

Steps to reproduce:
1. Install Accounting
2. Create a company with 'CHF' as currency and select it
3. Enable 'Multi-Currencies' in Settings > Accounting > Currencies
4. Go to Accounting > Configuration > Invoicing > Journals
5. Create a new journal of type 'Bank' and currency 'EUR'
6. Go to the Accounting Dashboard and import the CAMT file (go to the
three dots on the top left corner of the bank journal and select New >
Import Statement)

Solution:
With `entry_amount` = the amount in \<Ntry>/\<Amt> and `amount` is the
amount of the transaction (\<Ntry>/\<AmtDtls>/\<TxAmt>/\<Amt>), we check
 that `entry_amount = amount * rate` if not, we do the same check with
`rate = 1/rate`. If still not, we just use the `entry_amount`.

Problem:
The exchange rate in the CAMT053 file provided by banks is sometimes
inverted, the issue is more thouroughly explained here :
https://github.com/odoo/enterprise/pull/20419#issuecomment-925505605

opw-2689388

closes odoo/enterprise#25570

X-original-commit: df63265fe2b310ca20b92097383228e78ee1da58
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Guillaume Merlin (megu) <megu@odoo.com>