[IMP] account_reports: allows opening closing move of closed periods
At the moment, clicking "closing entry" for a closed period in a tax
report is showing a user error saying that the period is closed.
This change aim to improve that by instead displaying the closing move
itself if it exists.

Backport of https://github.com/odoo/enterprise/pull/24406

Task id #3184756

closes odoo/enterprise#37158

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>