[IMP] account_reports: make get_options_domain more robust when fiscal_position option is missing
Before, when 'fiscal_position' option wasn't set in the options dictionary, we entered the else and added [('move_id.fiscal_position_id', '=', False)] to the domain, which was wrong. Though this case doesn't happen when properly initializing options with a call to _get_options(), we had a case in Luxembourg where it wasn't done. Making this change makes the code more robust against any future misuse of this nature.

OPW 2751338

closes odoo/enterprise#24462

X-original-commit: 56237da869a00cdc5e71209bc595ee5b21155857
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Olivier Colson <oco@odoo.com>