[FIX] l10n_cl_edi_boletas: Traceback generating daily sales book
If you execute the scheduled action "1. Cron Job - Run SII Sales Book
Report Process", this will result in a traceback because we call
`decode` on a `Markup` object which doesn't have that method. This is
due to a change in `ir_qweb.py`'s `_render` implementation which now
returns a `Markup` object instead of a `bytes` one:
https://github.com/odoo/odoo/pull/68299 .

It looks like daily sales book was missed during the corresponding
refactoring.

closes odoo/enterprise#24463

X-original-commit: 8d5a3cc2805af299af86bc3f839a13ad579cd5fe
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Raf Geens <raf@odoo.com>