[FW][MOV] appointment: split test files
Purpose of this commit is to split test files, notably to get UI related
tests outside of standard tests. This allows to limit the usage of HttpCase
to some tests notably.

Basic usage of appointment test class is also done. This will be improve in
master when cleaning more in-depth appointment code flow.

Task-2631088 (Appointment: profile and add performance tests)

X-original-commit: odoo/odoo@c477626c9df146b4300d0e1f4b0a9b78611468cac477626c9df146b4300d0e1f4b0a9b78611468ca
Part-of: odoo/enterprise#24559