[IMP] hr_payroll: change new contracts in employee trends dashboard
Currently when 'new contracts' link is clicked in the employee trends board,
all contracts with state=draft are shown. This commit changes the name of the
link from 'new contracts' to 'new employees' and instead of showing all
contracts with state=draft, we show all employees with contracts that have been
running in the last 3 months and that have no past contracts.

Adds 'new contracts' warning showing all contracts with state=draft.

Change employee trends chart from line to bars, in order to keep the same type of
chart as the employer costs chart.

task-2759601