[FIX] l10n_mx_edi: Correct invoice CFDI for fixed taxes.
Some useful context:
In Mexico, there is only one fixed tax, and it appears very rarely.
It is called the 'IEPS', and applies to fossil fuels, e.g., gasoline,
diesel, fuel oil etc. A fixed amount is charged per liter / ton of
fuel.

(see https://www.infoautonomos.mx/obligaciones-fiscales/impuesto-especial-sobre-produccion-servicios-ieps/#IEPS_2020_Gasolina_y_diesel)

In the CFDI, this fixed tax must be reported as follows:
- `TipoFactor="Cuota"`
- `TasaOCuota="(tax amount / base amount)"`
- `Importe="(tax amount)"`

Indeed, the PAC performs a verification based on the `TipoFactor`,
`TasaOCuota` and `Importe` fields.

If the `TipoFactor` field is "Tasa", the PAC checks that the
`TasaOCuota` amount is one of several authorized values
(see this table: https://sf.ezoiccdn.com/ezoimgfmt/www.cfdi.org.mx/wp-content/uploads/2020/02/Tabla-de-Valores-de-Tasa-o-Cuota-1024x670.png?ezimgfmt=ng:webp/ngcb1
found on https://www.cfdi.org.mx/catalogos-de-cfdi/tasa-o-cuota-de-impuestos/)

If the `TipoFactor` field is "Cuota", the PAC rather checks that the
`TasaOCuota` amount is comprised between 0 and 43.770.

Moreover, the PAC checks that the `Importe` field is equal to the base
amount times the `TasaOCuota` field, rounded either up or down.

The issue:

At the moment, instead of setting
`TasaOCuota="(tax amount / base amount)"`,
we set `TasaOCuota="(tax.amount / 100)"`.

This is correct for percentage of price taxes, but not for fixed
taxes.

And this causes PAC validation errors as can be seen in this video,
taken on a fresh v15 database:
https://drive.google.com/file/d/1APaeAi--RUbGtPXgRcyG_1Fyy9xtmn34/view?usp=sharing

This is also the issue in opw-2724805

The solution:
Correctly set the `TasaOCuota` field when the
`tax.l10n_mx_tax_type` is 'Cuota'.

closes odoo/enterprise#25018

X-original-commit: 9b4cfabad89b17a66ccf3c01236237eb4a268971
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>