[FIX] l10n_mx_edi_stock: move line weight is not correct when using serial numbers
The Mexico Delivery Guide modifies the standard delivery slip to include the weight of each line. When Serial Numbers are printed on the Delivery slip, the weight of the related stock.move is a sum of all units while they are separated on each line - resulting in a increased weight. This commit adds a new computed field on stock move line, and uses this field on the report instead

closes odoo/enterprise#24618

X-original-commit: 42bd00f238127da2c7bacebece7dcb23217e10bb
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Ayob Habib (ayh) <ayh@odoo.com>