[IMP] *: Deterministic mock server sorting
This commit is the enterprise counterpart of a commit of the same name.

Said commit orders the records coming from the mock server the same
way they are ordered by the ORM, while also allowing it to support
multiple levels of orderby.

This has been done to increase consistency in the test suite and provide
a more accurate representation of how records would be returned by the
actual server.

Some tests performed their assertions based on this previous
undeterministic system and have been adapted, either by altering
the base setup or the assertions themselves.

closes odoo/enterprise#24509

Related: odoo/odoo#84718
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>