[FIX] sign: adjust signature size
In #19577 a method `adjustSignatureSize`
was added to adjust the signature size with the signature item box size keeping
the original ratio of the image. However, the approach that was being used
altered significantly the resolution of the signature image. This commit
introduces a new approach where we change the canvas size instead of changing
the image size.

closes odoo/enterprise#25081

X-original-commit: 12145593ba5c5c074be38af731a598bf3bc8f411
Signed-off-by: Arnaud Joset <arj@odoo.com>