[FIX] website_calendar: allow user to create appointment type
Before this commit user was not able to create appointment type
from website if user's timezone is not set as, appointment_tz
is a required field in 'calendar.appointment.type'.

This commit will allow user to create appointment typw
also if user's timezone is not set. For that we have
create_and_get_website_url method and assigning
tz to appointment_tz. For tz value we have used the
website_visitors timezone as once the visitor visits
the website, his timezone is stored in cookies.
So, we are using that tz here.

TaskId-2745881

closes odoo/enterprise#25329

X-original-commit: 6bc03dcd48f4e3d84283286286d2a7b8feaa890d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>