[IMP] l10n_ar_reports: add cases for liquido producto
We create liquido producto data on the fly and update the comparation
files to ensure that this type of documents TXT exportation is working
properly: TXT possitve and negative lines for the refund liquido
producto documents.

Also, Doing the changes for puchase book we found an error in the sales
vat book, were the credit notes were created in the wrong journal. We
fix the issue in the tests and update the comparation txt files with the
proper fix.

closes odoo/enterprise#36073

X-original-commit: e6bdb77486865625ac4ad3aa0bc8951d6cfe92de
Related: odoo/odoo#110480
Signed-off-by: Josse Colpaert <jco@odoo.com>