[IMP] planning,project_forecast: improve the shift for resource type material
Purpose of the commit is to improve the behaviour of the shift for
resource type material.

So in this commit did the below changes:
 - the resource type material shift should be published by default
and it can not be unpublished and do not show the publication_warning
on those shifts.
Also removed _get_fields_breaking_publication method as publication_warning
is compute field with dependent on all other fields.

task-2758779

closes odoo/enterprise#26566

X-original-commit: 56ccfbd64d6fb39613e6b2894be891752ffca03a
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>