[FIX] l10n_*_{reports,edi,check_printing}: always auto_install
Some modules are set to be auto installed, but with dependencies
preventing the desired flow.
For instance, `account_lu_reports` depends on `account_saft`, but as we
don't want that module to be auto installed, `account_lu_reports` wasn't
either. To fix that, we declare the module as `auto_install`ed based on
the main `l10n_**` modules and the base `account_*` modules only.

It is done even when not stricly necessary to give a guideline, as it
was an issue for several modules not correctly auto-installed.

closes odoo/enterprise#24576

Related: odoo/odoo#84975
Signed-off-by: Quentin De Paoli <qdp@odoo.com>