[FIX] web_studio: only accept png file for custom icons
A previous PR (see odoo/enterprise#9339) added a filtering on the
accepted files when uploading a custom app icon in Studio.

This commit goes further by restricting to PNG files only as the rest of
the codebase for app icons only properly handles this image format (see
"data:image/png;base64..." in the IconCreator upload handler).

As an example, when trying to upload a JPEG file, the icon won't be
displayed properly.

task-2766273

closes odoo/enterprise#24939

X-original-commit: c103173cc34eba57041101c33ff1bf11623ce72b
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans (app) <app@odoo.com>