[FIX] web_enterprise: Force proper numbering system for server dates
Enterprise counterpart of a commit of the same name.

Before this commit, the numbering system used by luxon to parse and
format dates was the one set by the current locale. This means that
any date sent to and received from the server would also be expected
to use the current locale and numbering system. The consequence of
this behavior is that any domain generated by arbitrary dates would
trigger a server error.

This commit forces the numbering system to 'latn' (browser default) for
dates used by the server.

closes odoo/enterprise#24822

X-original-commit: c42c718253f0f903bbe0194fcf4c0b974485a363
Related: odoo/odoo#85487
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>