[IMP] planning,project_timesheet_forecast{_sale}: improve planning analysis report
Improve the Planning analysis report by providing additional measures on:

the capacity and utilization of resources -> goal: easily identify
who is still available or who has been allocated too much work
the ratio of billable vs. non-billable hours assigned -> goal:
making sure that a resource is profitable and that the work he is assigned is well balanced
the time off taken by the employees -> goal: help the user
understand the distribution of hours
Because these measures are quite complex to implement, we will be creating a new
report for this task instead of relying on planning.slot.

See task-2742620

closes odoo/enterprise#24699

Related: odoo/upgrade#3288
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>