[IMP] hr_payroll, hr_payroll_account_sepa: improve batch view in dashboard
Changes the batch view by removing the date column, changing the batch name.
Also changes the sepa download button to prevent the column to be two-lined.
The idea is to increase the available row space to the batch name and reduce
the download icon space by changing the text to "SEPA" instead of the file name.

task-2759601