[IMP] planning_holidays: update the slot duration and working day count
Reproducing Steps:
-Installed Planning App
-Visible working_days_count field in the slot
-Create Slot-(10-1-2021 to 14-1-2021) One Week(5-day working)
-Add public holiday(12-1-2021)
-If we check the working day of the slot, 5 days appear but 4 should appear

Prior to this commit:
The calculation of the working days not taking leaves into account
for working day slots.

After this commit:
Calling the method _compute_allocated_hours and _compute_working_days_count
methods for impacted shifts of planning.slot when leave is added, validated
and the date of leave is changed. So that allocated_hours and working_days_count
include holiday hours.

task-2731462

closes odoo/enterprise#24619

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>