[FIX] hr_payroll: print payslip with non Latin characters in filenames
There is an error when we are trying to print a payslip with
an employee with non Latin-1 characters in his name. Because the name
of the employee is written in the filename of the payslip.

Step to reproduce:
- Install the hr_payroll app
- create an employee with non-Latin characters in is name
- go to the Payslips of this employee
- create a new payslip for this employee
- and try to print it

Current behavior:
It's open a new tab and throw a `502 Bad Gateway` error because
the treads for printing the payslips crashes.

Behavior after this commit:
We are now using the function `content_disposition` to use the
percent-encoding defined here: https://www.rfc-editor.org/rfc/rfc3986

opw-2748662

closes odoo/enterprise#24855

X-original-commit: a1c58ff74e6f399e32bf835c4c4a5c2c48cb826e
Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>