[FIX] industry_fsm : fix showing all users of database when grouping by responsible
Purpose
=======

When working with multi-company, on **Field Service** model, when
grouping by responsible, all users of the database are listed, even those who don't
belong to the current company (or selected companies) where the user is performing the action.

The issue is that when the users list is computed, there is no domain that applies a condition
on the company.

Specification
=============
To solve the issue a domain on comany_id was added when computing the users list.

opw-2744423

closes odoo/enterprise#24974

X-original-commit: 946f1324b004d8609007230a62e8fc8d77f1bbc9
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: abla001 <abla@odoo.com>