[FIX] account_reports: tax report: don't try to format lines for generic reports as if they were localized reports
The condition at the beginning of _format_lines_for_display was written with the intend to prevent running this formatting on the generic tax reports, but it was wrong. Indeed, the 'tax_report' option is also set when using a generic report, with a string value identifying it.

This only caused useless computation in 15.0, but crashed in master (future 15.3), because we were browsing 'generic', which used to return a single record in 15.0-15.2, but now returns a record set, with a record for each letter. This, in turn, was breaking the ensure_one() in account.tax.report's get_checks_to_perform().

closes odoo/enterprise#24693

X-original-commit: 17ac7c65049e167ee0c285454b23bcac0ddfa2d2
Signed-off-by: William André (wan) <wan@odoo.com>