[FIX] industry_fsm_sale: get sale order items of project
This commit is the following of the fix in community. In that commit,
a condition is added in the `project.project` to fetch the SOL linked to
the project is the project is not a fsm project. Indeed, the fsm project
is linked to a service product instead of a SOL.

part of task-2710808

X-original-commit: 9f3e0ffadbffe42e8ec309541a0069c35efc8da7
Part-of: odoo/enterprise#24918
Co-authored-by: Yannick Tivisse <yti@odoo.com>
Co-authored-by: Laurent Stukkens (LTU) <ltu@odoo.com>