[FIX] account_sepa: no BIC with Swiss QR-IBAN Accounts
Ensure BIC is provided on Debtor in pain.001 as the Swiss flavor
doesn't accept /DbtrAgt/FinInstnId/Othr tag.

Also, the creditor agent can be omitted where IBAN/QR-IBAN is sent in
the Creditor Account. Currently, providing NOTPROVIDED in /Othr/Id is
not compatible with some Swiss banks such as ZKB.

closes odoo/enterprise#24801

X-original-commit: b6535cd7ff0763acdccf3922ae298eb7e6933e8e
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>