[FIX] sign: image drop causes pdf rendering error
After a sign request is completed and you visualize the document,
it is possible to drag/drop the signature images. When doing that,
pdf.js tries to open the dropped image, causing an error. This commit
edits the css of the completed sign items to prevent drag and drop.

task-2746307

closes odoo/enterprise#24707

X-original-commit: 4ce48688d45df5911fee6020e3310086711d1dcf
Signed-off-by: Arnaud Joset <arj@odoo.com>