[IMP] sign: handles incorrect file upload in sign
In sign it is possible to upload multiple files, however if one of them can't
be read, an error is raised and it's not possible to know which uploads failed.
This commit changes the way file upload is done, in order to allow clearer
messages. Single and multiple file upload is done using the same method call and
if there are failing PDFs, a message is returned to the client showing which
files failed. The choice of returning the message instead of raising an error
was intended, as we don't want to prevent further processing of the successful
files.

task-2697096

closes odoo/enterprise#23554

Signed-off-by: Arnaud Joset <arj@odoo.com>