[IMP] website_helpdesk_livechat: allow creating tickets within the chatbot
PURPOSE

This commits adds a special step_type on the chatbot.script.steps that allows
creating a helpdesk.ticket when reaching it.

SPECS

In addition, we let the end user configure in which helpdesk.team that ticket
will land upon being created, making it possible to sort your tickets based on
previous questions.
e.g: "In which country do you live"
And based on the answer, you assign to a different helpdesk.team.

This commit comes with the necessary business code to make this feature work
as well as some extra testing methods that ensure it is correctly implemented.

Task-2030386

closes odoo/enterprise#24894

Related: odoo/odoo#84000
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>