[FIX] sale_ubscription_dasnboard: default date range of salesmen report was displaying wrong data
Before this commit, when the user opened the salesmen dashboard, the default date filter was set to 'last month' but the data from current month were displayed.
When the filter was manually changed, the correct data were displayed. This commit esure that the data of the previous month are displayed when opening the dashboard with default value.

closes odoo/enterprise#25071

X-original-commit: 038267514b6afda07b4ddf4b090f2123c1759a9c
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>