[FIX] timer: Fix perf issues on timer computation
Before this commit, the _compute_user_timer_id method made a search to
get the timer of the current user for each record into the `self`. By
doing that, a query is made for each record, if more records are
contained in the self, more longer the time will be.

This commit changes this method to made a `read_group` to search all
timer linked to the current user for all records contained in `self`.
With this commit, the execution time for this method is reduced from 21
to 0.6 ms.

X-original-commit: 76016bc5264e0424e38405ef802e36bbf026e988
Part-of: odoo/enterprise#24918
Co-authored-by: Laurent Stukkens (LTU) <ltu@odoo.com>
Co-authored-by: Xavier BOL <xbo@odoo.com>