[ADD] [website_]appointment[_crm|_hr]: link visitor to lead
Bridge the gap between the web visitors and the lead created when they book a
 demo so that we link all data from A to Z.

Specification:

When opportunities are created from an appointment:
- add the language and the country to the opportunity
- link the visitor to the opportunity so that pages browsed by the user are
associated with the opportunity (which is useful to prepare the meeting)

We have also added the page /appointment as tracked.

Technical note:
- Although it can be done in a few lines, we have decided to create a new
module because the task depends on 3 modules: appointment, crm and website
and there were no module with such dependency.
- the page /appointment is rendered either as appointments_list_layout template
or appointments_cards_layout template depending on an option, so we have marked
both templates with the attribute track="1".
- the tracking of new pages has increased the query count in some test

Task-2778259

closes odoo/enterprise#24791

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>