[FIX] account_winbooks_import: refactor import methods
Use record generators to avoid taking up too much RAM when loading a
file, and work only with disk memory.
This should allow importing bigger files following the fiduciaries
requirements.

This reduces significantly the impact on the RAM as we are only loading
one file at the time.

At the same time, we are also dropping the manual management of the DBF
library since it is added in the python external dependencies in the
manifest.

task-2787480

X-original-commit: 82fc5297902f27752eba08522f9cfcb408cb3c77
Part-of: odoo/enterprise#25449