[FIX] industry_fsm_sale_report: fix unit test without demo data
Before this commit, when the industry_fsm_sale_report module is
installed without demo data, no journal could be defined in the company
used for the test and so the test failed because of no journal is
defined in the current company.

This commit uses the common class defined in the `industry_fsm_sale`
module to used the journal set by the common class defined in
`account` module.

closes odoo/enterprise#24913

X-original-commit: 4fa5ab1e7a12ee5455da758e294c06ba05f65233
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Xavier <xbo@odoo.com>