[FIX] planning,industry_fsm: fix display on mobile
Description of the issue/feature this PR addresses:

Steps to reproduce the issues:

Issues found using a mobile device only.

On the planning module, go to the gantt view and select a shift with a long name.
The title of the modal that appears will be displayed in three separate lines, and
not all the information is visible.

On the industry fsm module, since 14.0, the worksheet info when signing a document is now
correctly structured.

task-2755200

--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr

closes odoo/enterprise#26588

X-original-commit: 8aa59cdf25172c628162dd5952a98aebf30ff363
Related: odoo/odoo#89630
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>