[FW][PRF] appointment: do not block appointment with declined meetings
Some routes still use ``calendar_verify_availability`` method to check if
a partner is available for booking. It is not used anymore in slots generation
process but still in some final steps of booking.

In this commit we copy the new behavior and now check that attendee has
not declined the event before considering the slot to be already taken.

Task-2766574 (Appointment: improve performance)

X-original-commit: odoo/enterprise@de89dbe492e391f0f1f03587798eff582c07735e
Part-of: odoo/enterprise#25002