[FIX] {timesheet,web}_grid: fix performance issue on big database
When calling read_grid_grouped, a call to `_build_grid` is performed
for each `section`.

As the override of `_build_grid` in `timesheet_grid` is performing
database accesses, we potentially have (depending on the number of
sections) an explosion of database accesses.

The `hr_timesheet.timesheet_action_all` action, once `timesheet_grid`
is installed, is calling `_build_grid` for each employees that has
records for the selected period.

In order to solve that issue, this commit prevents increasing the
number of queries by triggering only once the group_expand for
all the grids generated by the `_build_grid` calls.

For a database where the action is returning around 500 employees:
- It was previously triggering about 1200 queries and was taking
  227 seconds.
- It is now triggering about 120 queries and runs in about 1,5 second.

closes odoo/enterprise#25203

X-original-commit: bda5bf401ac00bd4f56d65901475095113bd12ac
Signed-off-by: Xavier <xbo@odoo.com>