[FIX] helpdesk_mail_plugin: trigger the acknowledgement email
Bug
===
When we create a ticket from the Outlook / Gmail addin, the
acknowledgement email is not sent.

Technical
=========
The email is sent on a "stage" change (see `_track_template`). But in
mail thread, we do not care about value coming from the default_get.

So if the stage come from the default_get, the value is just ignored.

As all fixes in mail thread will add additional SQL queries, we just
fix it here (as nobody complained about that) and we give manually
the team (so the stage can be set in the create method of ticket
and the acknowledgement email is sent).

Task-2782150

closes odoo/enterprise#26718

X-original-commit: 4592f48612add464bc642c2064913a5547cbe043
Related: odoo/odoo#89965
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>