[FIX] timesheet_grid : register ts in multi-company environment
Steps :
Install project and employee.
Create 3 company and an employee for you user in 2 of them.
Switch to the one without employee.
Create a new project and task.
Register a timesheet.

Issue :
The timesheet has no employee.

Fix :
Display start button and set an employee on the ts
only if user has an employee in the company
or has only one employee for all companies.

Related : odoo/odoo#85780
opw-2745733

closes odoo/enterprise#25145

X-original-commit: f772f45e227279a45c405e9dbf9be809ce43d59d
Related: odoo/odoo#86135
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>