[FIX] l10n_be_reports: prefix phone number in xml
Steps to reproduce:
1) Create the system parameter 'l10n_be_reports.xml_export_representative_*'. To do so, first, ensure that you are on debug mode then, go to the application ‘Setting’, then click on ‘Technical > Parameter > System parameter’. Create a new parameter (l10n_be_reports.xml_export_representative_*) and replace the * by the ID of your company and define the value of the parameter as the Partner ID to be used as a proxy (ID of the accounting company).
For example: if the ID of your company is 4 and the ID of the fiduciary is 12, you will create the following parameter l10n_be_reports.xml_export_representative_4 with 12 as the value. More information in the following video: https://drive.google.com/file/d/1FVKqVXYyYa4VbBGZEomDFOBdt3_ZETsj/view
2) Ensure that the Partner has a phone number
3) Export XML for report (Partner VAT Listing)

Issue:
The file contains phone numbers with spaces => NOK

Cause:
Partners and companies numbers are sanitised automatically with `phonenumbers`. And those numbers have spaces in their format which is not accepted according to the specification https://finances.belgium.be/fr/E-services/Intervat/documentation-technique.
In the file `commontypes_v1.xsd`

Solution:
Extract the pattern accepted by intervat; That is, phone numbers with no spaces. The library `phonenumbers` has not been used to avoid complexity in case of the library is not installed.

OPW-2454136

closes odoo/enterprise#25725

X-original-commit: 0010624a09e74f6d0022766dd4480885c4ecd6d1
Signed-off-by: Olivier Colson <oco@odoo.com>
Signed-off-by: yosa-odoo <yosa@odoo.com>