[FIX] l10n_es_reports: adding rights to accountant group
How to reproduce the bug ?

- install the accounting and l10n_es modules
- login as a non-administrator user
- open the accounting app, then to go to Reporting > Tax Report (Mod 303)

What is the bug ?

As a non administrator user, you get an access error if you try generate
an accounting report for a spanish company with the spanish accounting module.
The error says that the user is not allowed to modify 'Companies'
(res.company) records.

opw-2767235

closes odoo/enterprise#25050

X-original-commit: b8c054a79bd92bef2f929de52753e34d12ba17c1
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>