[FW][PRF] appointment_hr: compute working hours in batch based on calendar
PURPOSE

Fetch information for employee availability in batch instead of fetching it
for each slot to check. This decreases a lot required queries.

SPECIFICATIONS

In this commit we benefits from ``resource`` module optimization to do the
working hours computation per calendar instead of per employee. When having
several employees linked to an appointment chances are high they actually
use the same base working calendar. This means we gain a lot of computation
by skipping computing several times the same working hours.

Leaves are taken into consideration based on resources as method is planned
to work in batch based on a resource record set. As leaves are attached to
the employee's resource they are correctly computed by resource module.

Task-2766574 (Appointment: improve performance)

X-original-commit: odoo/enterprise@5826c6239b12949df831f52da0daacdae45db258
Part-of: odoo/enterprise#25002