[FW][PRF] appointment(_hr): compute 'custom' availability in batch
PURPOSE

Fetch information for employee availability in batch instead of fetching it
for each slot to check. This decreases a lot required queries.

SPECIFICATIONS

Currently for 'custom' appointment types a pre-check for employee availability
is done directly in ``_slots_generate`` in addition to some check already
done in ``_slots_available``. An employee is set directly during generation
and computation done in availability check is somehow lost.

In this commit we remove the custom check done in generation and let custom
appointment types share the same code flow as other types. Everything is now
cleanly done in availability check.

We still have to manually skip the working hours computation as 'custom' types
do not check against calendar. Indeed we consider that custom hours are meant
to be prepared by sales people and this use case may be done outside of
working hours.

Task-2766574 (Appointment: improve performance)

X-original-commit: odoo/enterprise@de32b8f69c46039082d717c3401ea384b9f1aaf0
Part-of: odoo/enterprise#25002