[FW][IMP] appointment_hr: add performance test with warmup
See impact of cache in a given tests. Previous tests are done cold to try
to optimize as much as possible. But knowing impact of cache is also
interesting. Warmup will be added in more tests once main work about
performances in appointment is done.

A new test browsing /calendar/ is added on a type involving a lot of
work hours computation, allowing to better check improvements.

Task-2631088 (Appointment: profile and add performance tests)

X-original-commit: odoo/enterprise@1866db9f48f6cf559cc8a01c92939239b77ed4a6
Part-of: odoo/enterprise#25002