[FIX] helpdesk: fix the wrong value in sla failed filter
Before this commit, tickets with two SLAs failed that are in a closing stage
does not dislay under the sla_failed filter due to invalid computation of
sla_reached_late.

So this commit fixes the issue by adding the condition
for sla_reached_late so it will display the failed ticket in failed filter.

task-2784885

closes odoo/enterprise#27249

X-original-commit: 3438f11588daaf06ec36e092f553cd5ff6fa3d7c
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>