[FIX] l10n_lu_reports: wrong vat numbers for xml export
How to generate the bug ?

- install accounting and l10n_lu_reports
- in settings > general settings > companies, make sure that your lux
company has an "accouting firm" that has a different vat number than the
company.
- go to accounting app and open the 'Profit and Loss (abridged)' report
- generate a xml file by clicking on xml.

What is the bug ?

Whe you generate a xml file for some accounting reports for a Luxembourg
company, the vat numbers are wrong. Effectively, in the Declarations
section, the vat number should be the one of the company instead of the
the one of the accounting firm.

opw-2774820

closes odoo/enterprise#25850

X-original-commit: f54180e54262334e655a8a0c461cfbea5754c4a1
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>