[REF] documents_spreadsheet*: revamp list insertion
The purpose of this commit is to continue the move to not-scalable
commands to Core plugin instead of UI plugin.

The current implementation of ADD_ODOO_LIST is, like the old ADD_PIVOT,
done in a UI plugin, which dispatchs a tons of core commands. So, this
commit introduces two new core commands: INSERT_ODOO_LIST and
RE_INSERT_ODOO_LIST.

This commit also takes advantages of the concepts introduces in the
previous commit (ORM proxy, model factory, ...).

Since CacheRPC is not used anymore, we can remove it, same for legacyORM.

Task-id 2735400

Part-of: odoo/enterprise#24567