[FIX] l10n_ar_edi: HINT msg code error 10154
When the user gets the AFIP error code  10154 it indicates that the user
is trying to validate a CN/DN in AFIP but the value in the "FCE: Is
cancellation" field is wrongly set.

Depending on the conditions of the original document (has been or not
been rejected by the receptor) we tell the user to set True/False the
"FCE: Is cancellation". The problem is the recommendation message was
switched.  The proper message is:

* Origin Doc has been rejected then the NC should be marked as FCE Is
  Cancellation
* Origin Doc has NOT been rejected then the NC should **NOT** be marked
  as FCE Is Cancellation

closes odoo/enterprise#25004

closes odoo/enterprise#26054

X-original-commit: a02c28d6bf93316152013dfefb1629a12410565a
Signed-off-by: Josse Colpaert <jco@odoo.com>