[FIX] account_asset: copy asset type when duplicating assets
Currently, duplicating an asset can raise an error when saving it, as the asset type is not set.
Also, if an asset is opened from the Depreciation schedule view, there is no asset type in the context.
Asset type is computed from the context and thus in this case the asset cannot be saved.
Should be:
- Copy the asset type when duplicating to avoid the error.
- Add the asset type in the context when opening an asset from Depreciation Schedule view.

Task# 2781018

closes odoo/enterprise#27280

X-original-commit: 1045644d6c39594af38d50e9dd92e72fb177ed5b
Signed-off-by: William André (wan) <wan@odoo.com>