[FIX] Fix tax computations for DateV export with negative amounts
Before this commit:
Amounts for a line of an invoice were computed from the base amount and the taxes. There were problems when negative amount on a line were given (absolute value was used in the computation).

After this commit:
The amount is taken directly from the price_total which does the computation for us. We just take care of other types of entries like before.
Negative amounts on a line could happen after the fix, but it is corrected (by changing the debit/credit side).

A lot of code has been removed, as all the computation isn't needed anymore.

task-2759596

X-original-commit: d092be6dfd11cf6a2f9705c5bb8efa1766f77598
Part-of: odoo/enterprise#25179