[IMP] hr_payroll: change notes behavior and layout
Notes with the dashboard payroll tag now can be edited and viewed by all
payroll users that have access to the dashboard. This new behavior lets us
remove the former share mechanism for notes. However, we had to add a new
mechanism that shares the notes automatically with every payroll user. To
accomplish this it was necessary to overwrite the create and write methods
from note.note and from res.users.

Also, we removed the wizard for edition of the note in the dashboard. Now it is
possible to edit the name of the note and remove it if you are the creator of
the note.

task-2759601