[FIX] account_invoice_extract: fix wrong total after tax prediction
As the module `account_predictive_bills` can modify the taxes on
the line while the OCR is editing the total of the line, it could
lead to inconsistencies between what was detected and what was
encoded.

The better way to do this is to let it do its prediction before
modifying the total of the line.

closes odoo/enterprise#25266

X-original-commit: 67fc830446f9498b250cc7c03bef2a8335d153dd
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>