[FIX] project_enterprise: fix the multiple warning display in same alert
Before this commit, when there is more then 1 warning added on the same
alert then on any single warning condition all the warning are displayed.
For Ex: During the time off of an employee, the task in conflict warning is
displayed if there are no conflicts in Project.task of the gantt view

So this fix sets the attribute of the task button so that
when an employee takes time off, the task in conflict warning is no displayed
if there are no conflicts.

task-2774812

X-original-commit: 97716939b0cc6a7de01624250c1084e08c437812
Part-of: odoo/enterprise#27592