[IMP] industry_fsm_sale: add additional sales orders for profitability
Before this commit, when the "Extra Quotations" feature is enabled in
Field Service app, all projects could enable that feature, it means the
the additional quotations added with that feature should be taken into
account in the calculation of the project profitability.
However, it is not the case, only the SOLs linked to the project,
tasks, employee mappings and/or timesheets related to the project are
taken into account.

This commit adds the additional quotations added via that feature only
if they are a Sale Order. Because it is only when the order is made the
revenues could arrive. Indeed, if the quotation is not validated by the
customer then no revenue will be earned and no product in that quotation
will be delevered.

task-2710808