[ADD] test_sale_subscription: create historical data for subscription dashboards
Before this commit, demo subscription were created, no subscription invoices were generated and posted. This lack of meaningful records lead to empty dashboard.
This module creates subscriptions in the past in python and generated invoices for each periods, starting 12 months before the date of the module installation.

sale_subscription invoices are generated by comparing the next_invoice_date to today date, mail_tracking_values are generated when MRR changes.
It would be possible to generate these records in XML but they should be adapted each time changes are performed on sale subscription kpi generation.
Doing it in python allows the data to be always representative to the sale-subscription module state.

taskid: 2568007